requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
openai==0.27.0
